+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - GATT Client: Device Information Service Client
 - GATT Client: HID-over-GATT (HOG) Client, Report and Boot Host
 - GATT Client: Scan Parameters Service Client 
-- GATT Server: Scan Parameters Service Server 
+- GATT Server: Scan Parameters Service Server
 - GAP: support scan page configuration with `gap_set_page_scan_activity` and `gap_set_page_scan_type`
 - GAP: support sniff subrating with `gap_sniff_subrating_configure
 - GAP: support QoS setup with `gap_qos_set`
@@ -72,6 +72,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - HFP: provide acl_handle in events to identify connection
 - HCI Transport: extract convenience function declaration for h4, h5, em9304_spi, and usb into separate hci_transport_{type}.h
 - GATT Client: Use ATT_READ_REQUEST for first blob of Read Long Characteristic and Read Long Characteristic Descriptor
+- GATT Server: Allow ATT Read Callback to return custom ATT Error Code
 - HID: Move `src/classic/hid.h` into `src` and prefix with `btstack_` to use it with BLE and avoid name clashes
 
 ## Release v1.3.2